optimize version control by minimizing duplicate code
